Split workspace composer engine and wallet preflight
